+++ test/csp.js
@@ -1,10 +1,10 @@
 /*global describe:false, it:false */
 'use strict';
 
-var request = require('supertest');
-var assert = require('assert');
-var lusca = require('../index');
-var mock = require('./mocks/app');
+const request = require('supertest');
+const assert = require('assert');
+const lusca = require('../index');
+const mock = require('./mocks/app');
 
 describe('CSP', function () {
   it('method', function () {
@@ -12,12 +12,8 @@ describe('CSP', function () {
   });
 
   it('header (report)', function (done) {
-    var config = require('./mocks/config/cspReport');
-    var app = mock({ csp: config });
-
-    app.get('/', function* () {
-      this.body = 'hello';
-    });
+    const config = require('./mocks/config/cspReport');
+    const app = mock({ csp: config });
 
     request(app.listen())
     .get('/')
@@ -28,12 +24,8 @@ describe('CSP', function () {
 
 
   it('header (enforce)', function (done) {
-    var config = require('./mocks/config/cspEnforce');
-    var app = mock({ csp: config });
-
-    app.get('/', function* () {
-      this.body = 'hello';
-    });
+    const config = require('./mocks/config/cspEnforce');
+    const app = mock({ csp: config });
 
     request(app.listen())
     .get('/')
@@ -43,14 +35,10 @@ describe('CSP', function () {
   });
 
   it('string config', function (done) {
-    var app = mock({
+    const app = mock({
       csp: {
         policy: 'default-src *'
       }
-    });
-
-    app.get('/', function (req, res) {
-      res.status(200).end();
     });
 
     request(app.listen())
@@ -60,14 +48,10 @@ describe('CSP', function () {
   });
 
   it('array config', function (done) {
-    var app = mock({
+    const app = mock({
       csp: {
         policy: ['default-src *', 'img-src *']
       }
-    });
-
-    app.get('/', function* () {
-      this.body = 'hello';
     });
 
     request(app.listen())
@@ -77,16 +61,13 @@ describe('CSP', function () {
   });
 
   it('nested config', function (done) {
-    var app = mock({
+    const app = mock({
       csp: {
         policy: [
         { 'default-src': '*' },
         'img-src *'
         ]
       }
-    });
-    app.get('/', function* () {
-      this.body = 'hello';
     });
 
     request(app.listen())