+++ test/hsts.js
@@ -1,9 +1,9 @@
 'use strict';
 
-var request = require('supertest');
-var assert = require('assert');
-var lusca = require('../index');
-var mock = require('./mocks/app');
+const request = require('supertest');
+const assert = require('assert');
+const lusca = require('../index');
+const mock = require('./mocks/app');
 
 describe('HSTS', function () {
   it('method', function () {
@@ -17,12 +17,8 @@ describe('HSTS', function () {
   });
 
   it('header (maxAge)', function (done) {
-    var config = { hsts: { maxAge: 31536000 } };
-    var app = mock(config);
-
-    app.get('/', function* () {
-      this.body = 'hello';
-    });
+    const config = { hsts: { maxAge: 31536000 } };
+    const app = mock(config);
 
     request(app.listen())
     .get('/')
@@ -32,12 +28,8 @@ describe('HSTS', function () {
   });
 
   it('header (maxAge 0)', function (done) {
-    var config = { hsts: { maxAge: 0 } };
-    var app = mock(config);
-
-    app.get('/', function* () {
-      this.body = 'hello';
-    });
+    const config = { hsts: { maxAge: 0 } };
+    const app = mock(config);
 
     request(app.listen())
     .get('/')
@@ -47,12 +39,8 @@ describe('HSTS', function () {
   });
 
   it('hsts = number', function (done) {
-    var config = { hsts: 31536000 };
-    var app = mock(config);
-
-    app.get('/', function* () {
-      this.body = 'hello';
-    });
+    const config = { hsts: 31536000 };
+    const app = mock(config);
 
     request(app.listen())
     .get('/')
@@ -62,12 +50,8 @@ describe('HSTS', function () {
   });
 
   it('header (maxAge; includeSubDomains)', function (done) {
-    var config = { hsts: { maxAge: 31536000, includeSubDomains: true } };
-    var app = mock(config);
-
-    app.get('/', function* () {
-      this.body = 'hello';
-    });
+    const config = { hsts: { maxAge: 31536000, includeSubDomains: true } };
+    const app = mock(config);
 
     request(app.listen())
     .get('/')
@@ -77,12 +61,8 @@ describe('HSTS', function () {
   });
 
   it('header (maxAge; includeSubDomains; preload)', function (done) {
-    var config = { hsts: { maxAge: 31536000, includeSubDomains: true, preload: true } };
-    var app = mock(config);
-
-    app.get('/', function* () {
-      this.body = 'hello';
-    });
+    const config = { hsts: { maxAge: 31536000, includeSubDomains: true, preload: true } };
+    const app = mock(config);
 
     request(app.listen())
     .get('/')