+++ test/xssprotection.js
@@ -1,9 +1,9 @@
 'use strict';
 
-var request = require('supertest');
-var assert = require('assert');
-var lusca = require('../index');
-var mock = require('./mocks/app');
+const request = require('supertest');
+const assert = require('assert');
+const lusca = require('../index');
+const mock = require('./mocks/app');
 
 describe('xssProtection', function () {
   it('method', function () {
@@ -11,8 +11,8 @@ describe('xssProtection', function () {
   });
 
   it('xssProtection = 1', function (done) {
-    var config = { xssProtection: 1 };
-    var app = mock(config);
+    const config = { xssProtection: 1 };
+    const app = mock(config);
 
     request(app.listen())
     .get('/')
@@ -21,8 +21,8 @@ describe('xssProtection', function () {
   });
 
   it('header (enabled)', function (done) {
-    var config = { xssProtection: true };
-    var app = mock(config);
+    const config = { xssProtection: true };
+    const app = mock(config);
 
     request(app.listen())
     .get('/')
@@ -31,8 +31,8 @@ describe('xssProtection', function () {
   });
 
   it('header (enabled; custom mode)', function (done) {
-    var config = { xssProtection: { enabled: 1, mode: 'foo' } };
-    var app = mock(config);
+    const config = { xssProtection: { enabled: 1, mode: 'foo' } };
+    const app = mock(config);
 
     request(app.listen())
     .get('/')
@@ -41,8 +41,8 @@ describe('xssProtection', function () {
   });
 
   it('header (enabled is boolean; custom mode)', function (done) {
-    var config = { xssProtection: { enabled: true } };
-    var app = mock(config);
+    const config = { xssProtection: { enabled: true } };
+    const app = mock(config);
 
     request(app.listen())
     .get('/')
@@ -51,8 +51,8 @@ describe('xssProtection', function () {
   });
 
   it('header (!enabled)', function (done) {
-    var config = { xssProtection: { enabled: 0 } };
-    var app = mock(config);
+    const config = { xssProtection: { enabled: 0 } };
+    const app = mock(config);
 
     request(app.listen())
     .get('/')